# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Paste your JWT token here in the format: Bearer <JWT_TOKEN>"
    }
  },
  "paths": {
...
    "/admin/login": {
      "post": {
        "tags": [
          "Login Admin"
        ],
        "description": "Logs in Admins.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets the list of all available products.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets a specific product.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets the list of all available categories.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets a specific category.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets the list of all available brands.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/brands/{id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets a specific brand.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/search/": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Searches for a product either by fully or partially entering the name or by entering the category or brand or all three together.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Gets the list of all available orders.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Gets a specific order.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Admin": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "1234"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
...
    "Search": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "iPhone or pho"
        },
        "category": {
          "type": "string",
          "example": "Phones"
        },
        "brand": {
          "type": "string",
          "example": "Apple"
        }
      },
      "required": [
        "name",
        "category",
        "brand"
      ]
    },
...
  }
}